move tests to examples; fix broken tests after renaming;
