+++ examples/botkit-slack/tests/sample_echo_Spec.js
@@ -1,11 +1,11 @@
 'use strict';
 const assert = require('assert');
-const {Botmock, SlackApiMock} = require('../../../lib');
+const {BotMock, SlackApiMock} = require('../../../lib');
 
 const {SlackAdapter, SlackMessageTypeMiddleware, SlackEventMiddleware} = require('botbuilder-adapter-slack');
 const fileBeingTested = require('../features/sample_echo');
 
-describe('sample_echo file tests', () => {
+describe('sample_echo file general-slack', () => {
     beforeEach(() => {
         this.userInfo = {
             slackId: 'user123',
@@ -21,7 +21,7 @@ describe('sample_echo file tests', () => {
         adapter.use(new SlackEventMiddleware());
         adapter.use(new SlackMessageTypeMiddleware());
 
-        this.controller = new Botmock({
+        this.controller = new BotMock({
             adapter: adapter,
         });
