+++ examples/general-slack/slashCommand/slashCommand.js
