+++ examples/general-slack/slashCommand/slashCommandSpec.js
@@ -1,11 +1,11 @@
 'use strict';
 const assert = require('assert');
-const {Botmock, SlackApiMock} = require('../../lib');
+const {BotMock, SlackApiMock} = require('../../../lib');
 
-const {SlackAdapter, SlackMessageTypeMiddleware, SlackEventMiddleware} = require('botbuilder-adapter-slack');
+const {SlackAdapter, SlackMessageTypeMiddleware, SlackEventMiddleware} = require('botbuilder-adapter-slack/lib/index');
 const fileBeingTested = require('./slashCommand');
 
-describe('slash command tests', () => {
+describe('slash command general-slack', () => {
     beforeEach(() => {
         this.userInfo = {
             slackId: 'user123',
@@ -38,7 +38,7 @@ describe('slash command tests', () => {
         adapter.use(new SlackEventMiddleware());
         adapter.use(new SlackMessageTypeMiddleware());
 
-        this.controller = new Botmock({
+        this.controller = new BotMock({
             adapter: adapter,
             disable_webserver: true
         });
