+++ lib/BotMock.js
@@ -3,7 +3,7 @@ const {Botkit} = require('botkit');
 const {TurnContext} = require('botbuilder');
 const {setTimeoutAsync} = require('./utils');
 
-class Botmock extends Botkit {
+class BotMock extends Botkit {
 	constructor(config) {
 		// try to disable botkit warning if config has no disable_console property
 		if(!(config).hasOwnProperty('disable_console')){
@@ -96,4 +96,4 @@ class Botmock extends Botkit {
 	}
 }
 
-module.exports = Botmock;
+module.exports = BotMock;
