+++ lib/Slack/api.js
@@ -71,7 +71,7 @@ module.exports.defaultSlackEndpoints = {
 		user: {
 			name: 'test',
 			profile: {
-				email: 'tests@gmail.com'
+				email: 'general-slack@gmail.com'
 			}
 		}
 	},
